Remove stop function
I added a 'stop' function in anticipation of turning this into a daemon program. I attempted to do so using python-daemon, but was unsuccessful. I decided that I don't want to get into the weeds of daemonization right now, so I removed the 'stop' function. For now, if you want to use this program you will have to leave a terminal window open or detach it yourself.